refactor(ui): Remove unused prototype assets
