+++ Config/DefaultEngine.ini
@@ -215,8 +215,7 @@ UIScaleRule=ShortestSide
 CustomScalingRuleClass=None
 UIScaleCurve=(EditorCurveData=(Keys=((Time=480.000000,Value=0.444000),(Time=720.000000,Value=0.666000),(Time=1080.000000,Value=1.000000),(Time=8640.000000,Value=8.000000)),DefaultValue=340282346638528859811704183484516925440.000000,PreInfinityExtrap=RCCE_Constant,PostInfinityExtrap=RCCE_Constant),ExternalCurve=None)
 bAllowHighDPIInGameMode=False
-DesignScreenSize=(X=1920,Y=1080)
-bLoadWidgetsOnDedicatedServer=True
+
 bAuthorizeAutomaticWidgetVariableCreation=False
 CustomFontDPI=96
 FontDPIPreset=Standard
@@ -242,3 +241,5 @@ EditorStartupMap=/Game/Levels/L_FretteOpenWorld.L_FretteOpenWorld
 GameDefaultMap=/Game/Levels/L_FretteOpenWorld.L_FretteOpenWorld
 GlobalDefaultGameMode=/Game/Core/BP_FretteGameMode.BP_FretteGameMode_C
 
+
+[CoreRedirects]
