+++ Content/Core/BP_FrettePlayerCharacter.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:7245d7435d943566b637c1858bbb0d1b49ce75808a723f31d92b15391c54c232
-size 80101
+oid sha256:de10a7bab02ecbb9d55ad67bac305d1892818d633efe30c06ee44e599b57655a
+size 83776
