+++ Content/Core/BP_FrettePlayerController.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:4fbb8211bcf3f50e7250e5a05e78d253e2087564d0363c6368b6f95465917441
-size 30613
+oid sha256:b223e84aae333a89dad48582b0547d84c58c1231731915ced7bcbdeb8667d7fd
+size 30396
