--- Content/Core/DA_Skibidi.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:0891dbac25f7a75ed9bb2799cbca429b3e65cda672348bf5e0db5716b2f38828
-size 1683
