--- Content/Core/DA_Sword.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:2f521690d4dfe9ea450167c317bc25d7725a164928e0a7a9709ade8352827260
-size 1410
