--- Content/Core/SpecialTets.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:7a316d0938d05852cf54a12aa6672fe3ce1fe2e1189c94a8e6e6d976f723c9d5
-size 1137
