--- Content/UI/Test/AC_GarbageInventory.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:1e03879fe80278cdf8c98ae4ea37e995dc072b4f8342396ee2e590819e2333bd
-size 14150
