--- Content/UI/Test/AC_Test.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:4e11a2c5d3ead66dbc38db8f6c22a722468043fd41fb040a4a700a9b0227d506
-size 57408
