+++ Content/UI/Test/BP_RuntimeItem.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:1d2d244cffc9292577ec1228155d0646307e4b7b0cef9f76de2081f30ef8da13
-size 7483
+oid sha256:b7045ff68c2de146ea5324d9c86d2433ae6fc0b3fdf323dd3d3fa2046242407c
+size 7600
