--- Content/UI/Test/DA_InventoryItem.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:75d89ae1e5a0d489d6b3b99b24708bc307b3166b353e4268d42f36265f4e3395
-size 7710
