--- Content/UI/Test/VM_Test.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:786aacea900b01f274444a30760d245b9b9befd1c348d64a09e9b7ece61d7c04
-size 44566
