+++ Content/UI/Test/WBP_Inventory.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:58d621bf17c7c9f12dd2bb91415bda07edeeedf03607ae72a49d29cad6222f96
+size 31539
