+++ Content/UI/Test/WBP_InventoryItem.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:63dbdaf6d08354f407f5f2c916325dc93c2f61a30d32e3ae6f583f3e7228a1da
-size 36287
+oid sha256:4be8d93ec3ee1ed076a1acd507f2682d89d42332315d0864e9a3c17a9d399b8b
+size 36991
