--- Content/UI/Test/WBP_Test.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:129a52e2b154b431415511babcf60b926eba21c557b647d2d61fd3656e3f989c
-size 47320
