--- Content/UI/VM_Test.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:57df1e7fc32604853567dc91ff622616b4f9de02f7fdfd611835889f7d1f9375
-size 2100
