--- Content/UI/WBP_HUD.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:9b167c8532d4a577a31318b4a9bf9e9b84555a6a926201d522bb92da9639d7b4
-size 2220
