--- Content/UI/WBP_Inventory.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:c99135fb006bd592c66ac5836762f0563203c2475f5a65863fcc2b36b4a7c5ff
-size 30166
