+++ Frette.uproject.DotSettings.user
@@ -1,0 +1,3 @@
+﻿<wpf:ResourceDictionary xml:space="preserve" xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" xmlns:s="clr-namespace:System;assembly=mscorlib" xmlns:ss="urn:shemas-jetbrains-com:settings-storage-xaml" xmlns:wpf="http://schemas.microsoft.com/winfx/2006/xaml/presentation">
+	<s:String x:Key="/Default/CodeInspection/ExcludedFiles/FilesAndFoldersToSkip2/=7020124F_002D9FFC_002D4AC3_002D8F3D_002DAAB8E0240759_002Ff_003ARivermaxLib_002Ebuild_002Ecs_002Fl_003A_002E_002E_003F_002E_002E_003FUE_005F5_002E7_003FEngine_003FSource_003FThirdParty_003FNVIDIA_003FRivermax_003FRivermaxLib_002Ebuild_002Ecs/@EntryIndexedValue">ForceIncluded</s:String>
+	<s:String x:Key="/Default/CodeInspection/ExcludedFiles/FilesAndFoldersToSkip2/=77114B07_002DCE0E_002DDE3B_002D671D_002D7547DA422920_002Fd_003A1_002E30_002E2_002Fd_003Aasio_002Fd_003Adetail_002Ff_003Awinsock_005Finit_002Ehpp/@EntryIndexedValue">ForceIncluded</s:String></wpf:ResourceDictionary>
