+++ Source/Frette/Private/FrettePlayerCharacter.cpp
@@ -7,7 +7,7 @@ AFrettePlayerCharacter::AFrettePlayerCharacter()
 	Camera = CreateDefaultSubobject<UCameraComponent>(TEXT("Camera"));
 	Camera->SetupAttachment(GetCapsuleComponent());
 	
-	InventoryComponent = CreateDefaultSubobject<UInventoryComponent>(TEXT("Inventory"));
+	InventoryComponent = CreateDefaultSubobject<UFakeInventoryComponent>(TEXT("Inventory"));
 }
 
 void AFrettePlayerCharacter::DoPlayerMove(FVector2D MoveAxis)
