+++ Source/FretteUI/Private/FretteViewModel.h
@@ -2,7 +2,6 @@
 
 #include "CoreMinimal.h"
 #include "MVVMViewModelBase.h"
-#include "Misc/DataValidation.h"
 #include "FretteViewModel.generated.h"
 
 UCLASS(Abstract)
@@ -12,20 +11,5 @@ class FRETTEUI_API UFretteViewModel : public UMVVMViewModelBase
 	
 public:
 	virtual void Bind(APlayerController* OwningController) {}
-	
-private:
-	virtual EDataValidationResult IsDataValid(FDataValidationContext& Context) const override
-	{
-		const UClass* Class = GetClass();
-		
-		if (Class->IsInBlueprint() && !Class->IsFunctionImplementedInScript("Bind"))
-		{
-			const FText Error = FText::Format(INVTEXT("ViewModel '{0}' does not implement a Bind() event."), FText::FromString(GetName()));
-			Context.AddError(Error);
-			return EDataValidationResult::Invalid;
-		}
-		
-		return EDataValidationResult::Valid;
-	}
 };
